scanner-mobile/requirements.txt
Flask==3.1.2
mysql_connector_repackaged==0.3.1
Requests==2.32.5
python-dotenv
gunicorn
mysql-connector-python>=8.0
